requirements.txt
numpy==2.1.3
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==11.0.0
streamlit==1.40.1
tensorflow==2.17.1
tensorflow_intel==2.17.1
ultralytics==8.3.31
